% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEPC.R
\name{predict.ePCfit}
\alias{predict.ePCfit}
\title{Project New Data onto a Fitted Environmental PCA}
\usage{
\method{predict}{ePCfit}(object, newdata, ...)
}
\arguments{
\item{object}{An \code{"ePCfit"} from \code{\link{computeEPC}}.}

\item{newdata}{A data frame containing the same predictor columns.}

\item{...}{Unused.}
}
\value{
A matrix of ePC scores for \code{newdata}.
}
\description{
Project New Data onto a Fitted Environmental PCA
}
\examples{
data(durumUnits)
ep <- computeEPC(durumUnits[1:800, ])
newscores <- predict(ep, durumUnits[801:1060, ])
}
\seealso{
\code{\link{computeEPC}}
}
